+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Updated
   * AZ-273: Module now compatible terraform `v0.13+`
+Changed
+  * AZ-325: Upgrade keyvault module to latest version + Update CI
+  * AZ-326: Add missing keuvault's inputs
 
 # v3.1.1 - 2020-08-06